+++ src/constants/sess_keys.ts
@@ -1,0 +1,1 @@
+export const SESSION_NEWS_KEY = 'nsf_news'
