+++ src/loaders/index.ts
@@ -1,1 +1,5 @@
+export * from './loadData'
+export * from './facebookLoader'
 export * from './instagramLoader'
+export * from './newsLoader'
+export * from './meetLoader'
