+++ src/loaders/loadData.ts
@@ -1,65 +1,8 @@
-import { removeUndefinedFromArray } from '../utils/removeUndefinedFromArray'
-import { UPCOMING_MEETS, PREVIOUS_MEETS, NEWS } from '../views/Meets/constants'
+import { loadFBPosts } from './facebookLoader'
+import { loadIGPosts } from './instagramLoader'
+import _ from 'lodash'
 
-export interface IMeet {
-  title: string
-  description?: string
-  link: string
-  date: Date
-}
-export enum MeetType {
-  UPCOMING,
-  PREVIOUS,
-}
-/**
- * Cache meet data into session storage. .
- */
-export async function cacheMeets<T>(sessionKey: string, meetType: MeetType): Promise<T> {
-  if (window.sessionStorage.getItem(sessionKey)) {
-    switch (meetType) {
-      case MeetType.PREVIOUS:
-        return loadPreviousMeets(sessionKey) as T
-      case MeetType.UPCOMING:
-        return loadUpcommingMeets(sessionKey) as T
-    }
-  }
-
-  let meets
-  switch (meetType) {
-    case MeetType.PREVIOUS:
-      meets = await getPreviousMeets()
-      break
-    case MeetType.UPCOMING:
-      meets = await getUpcommingMeets()
-      break
-  }
-  window.sessionStorage.setItem(sessionKey, JSON.stringify(meets))
-  return meets as unknown as T
-}
-
-export async function cacheNews(sessionKey: string): Promise<INews[]> {
-  if (window.sessionStorage.getItem(sessionKey)) {
-    return JSON.parse(window.sessionStorage.getItem(sessionKey)!) as INews[]
-  }
-
-  const news = await getNews()
-  window.sessionStorage.setItem(sessionKey, JSON.stringify(news))
-  return news
-}
-
-function loadPreviousMeets(sessionKey: string) {
-  return JSON.parse(window.sessionStorage.getItem(sessionKey)!, (key, value) => {
-    return key === 'date' ? new Date(value) : value
-  })
-}
-
-function loadUpcommingMeets(sessionKey: string) {
-  return JSON.parse(window.sessionStorage.getItem(sessionKey)!, (key, value) => {
-    return key === ('startDate' || 'endDate') ? new Date(value) : value
-  })
-}
-
-enum DOMString {
+export enum DOMString {
   TEXT_HTML = 'text/html',
   TEXT_XML = 'text/xml',
   APP_XML = 'application/xml',
@@ -67,7 +10,7 @@ enum DOMString {
   IMAGE = 'image/svg+xml',
 }
 
-async function fetchDocument(endpoint: string, docType: DOMString): Promise<Document> {
+export async function fetchDocument(endpoint: string, docType: DOMString): Promise<Document> {
   return fetch(endpoint, {
     method: 'GET',
   })
@@ -76,101 +19,31 @@ async function fetchDocument(endpoint: string, docType: DOMString): Promise<Docu
       new window.DOMParser().parseFromString(str, docType as unknown as DOMParserSupportedType),
     )
 }
-export interface INews {
-  title: string | null
-  url?: string
-}
-export async function getNews(): Promise<INews[]> {
-  const news = await fetchDocument(NEWS, DOMString.TEXT_HTML)
-  const container = news.getElementsByClassName('slide-entry-wrap')[0]
-  const articles = Array.from(container.children)
-  const articletitles = []
-  const articlelinks = []
-  return articles
-    .map((article) => article.getElementsByClassName('slide-entry-title entry-title')[0])
-    .map((article) => {
-      return {
-        title: article.textContent,
-        url: article.innerHTML.match(/href="([^"]*)/)![1],
-      }
-    })
-}
 
-export async function getPreviousMeets(): Promise<IMeet[]> {
-  const meetData = await fetchDocument(PREVIOUS_MEETS, DOMString.APP_XML)
-
-  const channel = meetData.querySelector('channel')
-  if (!channel) {
-    return []
-  }
-  const meets = Array.from(channel.children).filter((node) => node.nodeName === 'item')
-
-  return removeUndefinedFromArray(
-    meets.map((meet) => {
-      const title = meet.querySelector('title')
-      const date = meet.querySelector('pubDate')
-      const description = meet.querySelector('description')
-      const link = meet.querySelector('link')
-      if (!title || !date || !description || !link) {
-        return undefined
-      }
-      return {
-        title: title.innerHTML,
-        date: new Date(date.innerHTML),
-        link: link.innerHTML,
-        description: description.innerHTML,
-      }
-    }),
-  )
-}
-
-async function fetchiCAL(endpoint: string): Promise<string> {
-  return fetch(endpoint, {
-    method: 'GET',
-  }).then((res) => res.text())
-}
-
-export interface IUpcommingMeet {
-  startDate: Date
-  endDate: Date
+export interface ICard {
+  id: string
+  src?: string
   title: string
-  weighIn: string
+  date: Date
+  text: string
   url: string
+  type?: string
 }
 
-export async function getUpcommingMeets() {
-  return (await fetchiCAL(UPCOMING_MEETS))
-    .split('BEGIN:VEVENT')
-    .slice(1)
-    .map((event: string) => {
-      const [start, end, title, weighIn] = event
-        .split(/\n/g)
-        .slice(1, -2)
-        .map((line) => line.split(/[A-Z][A-Z]*:/g)[1])
+export async function loadPosts(): Promise<ICard[]> {
+  if (window.sessionStorage.getItem('posts')) {
+    return JSON.parse(window.sessionStorage.getItem('posts')!, (k, v) => {
+      return k === 'date' ? new Date(v) : v
+    }) as ICard[]
+  }
 
-      return {
-        startDate: icalToDate(start),
-        endDate: icalToDate(end),
-        title: title.replace(/\r/g, ''),
-        weighIn,
-        url: `https://styrkeloft.no/terminliste/?page=terminliste&aar=${new Date().getFullYear()}&mnd=&k%5B%5D=13`,
-      } as IUpcommingMeet
-    })
-    .filter((event: IUpcommingMeet) => event.startDate >= new Date())
-}
+  const igposts = await loadIGPosts()
+  const fbposts = await loadFBPosts()
 
-/**
- * convert ical format to date object
- * Typicall ical format "20211022T000000"
- * Year: 2021
- * Month: 10 / October
- * Day: 22
- *
- */
-function icalToDate(ical: String): Date {
-  const YEAR = Number(ical.substring(0, 4))
-  const MONTH = Number(ical.substring(4, 6)) - 1 //Date starts at 0 = Jan and so on..
-  const DAY = Number(ical.substring(6, 8))
+  const output = _.uniqBy([...igposts, ...fbposts], 'text')
+  // TODO this is O(n²) complexity... would be better if we could ONLY query Facebook posts from the api to remove this duplicate removal funciton
 
-  return new Date(YEAR, MONTH, DAY)
+  const result = output.sort((a: ICard, b: ICard) => b.date.getTime() - a.date.getTime())
+  window.sessionStorage.setItem('posts', JSON.stringify(result))
+  return result
 }
